# Newid tweaks