--- a/src/MassTransit.Abstractions/NewId/NewId.cs
+++ b/src/MassTransit.Abstractions/NewId/NewId.cs
@@ -1,12 +1,12 @@
 namespace MassTransit
 {
     using System;
-    using System.Diagnostics;
     using System.Runtime.CompilerServices;
     using System.Threading;
     using NewIdFormatters;
     using NewIdProviders;
 #if NET6_0_OR_GREATER
+    using System.Diagnostics;
     using System.Runtime.Intrinsics.X86;
     using System.Runtime.Intrinsics;
     using System.Runtime.InteropServices;
@@ -51,7 +51,7 @@ namespace MassTransit
             if (bytes.Length != 16)
                 throw new ArgumentException("Exactly 16 bytes expected", nameof(bytes));
 
-            FromByteArray(bytes, out _a, out _b, out _c, out _d);
+            FromByteArray(bytes, out this);
         }
 
         public NewId(in string value)
@@ -59,11 +59,7 @@ namespace MassTransit
             if (string.IsNullOrEmpty(value))
                 throw new ArgumentException("must not be null or empty", nameof(value));
 
-            var guid = new Guid(value);
-
-            var bytes = guid.ToByteArray();
-
-            FromByteArray(bytes, out _a, out _b, out _c, out _d);
+            FromGuid(new Guid(value), out this);
         }
 
         public NewId(int a, int b, int c, int d)
@@ -286,42 +282,14 @@ namespace MassTransit
 
         public static NewId FromGuid(in Guid guid)
         {
-        #if NET6_0_OR_GREATER
-            if (Ssse3.IsSupported && BitConverter.IsLittleEndian)
-            {
-                Span<Guid> span = stackalloc Guid[1];
-                span[0] = guid;
-                Span<byte> byteSpan = MemoryMarshal.Cast<Guid, byte>(span);
-
-                FromByteSpan(byteSpan, out var newId);
-                return newId;
-            }
-        #endif
-
-            var bytes = guid.ToByteArray();
-            FromByteArray(bytes, out var a, out var b, out var c, out var d);
-
-            return new NewId(a, b, c, d);
+            FromGuid(guid, out var newId);
+            return newId;
         }
 
         public static NewId FromSequentialGuid(in Guid guid)
         {
-        #if NET6_0_OR_GREATER
-            if (Ssse3.IsSupported && BitConverter.IsLittleEndian)
-            {
-                Span<Guid> span = stackalloc Guid[1];
-                span[0] = guid;
-                Span<byte> byteSpan = MemoryMarshal.Cast<Guid, byte>(span);
-
-                FromSequentialByteSpan(byteSpan, out var newId);
-                return newId;
-            }
-        #endif
-
-            var bytes = guid.ToByteArray();
-            FromSequentialByteArray(bytes, out var a, out var b, out var c, out var d);
-
-            return new NewId(a, b, c, d);
+            FromSequentialByteArray(guid, out var newId);
+            return newId;
         }
 
         public byte[] ToByteArray()
@@ -535,47 +503,86 @@ namespace MassTransit
             return _getGenerator().NextSequentialGuid();
         }
 
-        static void FromByteArray(in byte[] bytes, out int a, out int b, out int c, out int d)
-        {
-            a = (bytes[10] << 24) | (bytes[11] << 16) | (bytes[12] << 8) | bytes[13];
-            b = (bytes[14] << 24) | (bytes[15] << 16) | (bytes[8] << 8) | bytes[9];
-            c = (bytes[7] << 24) | (bytes[6] << 16) | (bytes[5] << 8) | bytes[4];
-            d = (bytes[3] << 24) | (bytes[2] << 16) | (bytes[0] << 8) | bytes[1];
-        }
-
-    #if NET6_0_OR_GREATER
+    #if NET6_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        static void FromByteSpan(ReadOnlySpan<byte> span, out NewId newId)
+        static void FromGuid(in Guid guid, out NewId newId)
         {
-            Debug.Assert(span.Length == Unsafe.SizeOf<Guid>());
-            Debug.Assert(Ssse3.IsSupported && BitConverter.IsLittleEndian);
-
-            var vector = MemoryMarshal.Read<Vector128<byte>>(span);
-            var shuffle = Vector128.Create((byte)13, 12, 11, 10, 9, 8, 15, 14, 4, 5, 6, 7, 1, 0, 2, 3);
-            Vector128<byte> result = Ssse3.Shuffle(vector, shuffle);
-            newId = Unsafe.As<Vector128<byte>, NewId>(ref result);
-        }
-    #endif
-
-        static void FromSequentialByteArray(in byte[] bytes, out int a, out int b, out int c, out int d)
-        {
-            a = (bytes[3] << 24) | (bytes[2] << 16) | (bytes[1] << 8) | bytes[0];
-            b = (bytes[5] << 24) | (bytes[4] << 16) | (bytes[7] << 8) | bytes[6];
-            c = (bytes[8] << 24) | (bytes[9] << 16) | (bytes[10] << 8) | bytes[11];
-            d = (bytes[12] << 24) | (bytes[13] << 16) | (bytes[14] << 8) | bytes[15];
+            Span<byte> bytes = stackalloc byte[16];
+            guid.TryWriteBytes(bytes);
+            FromByteArray(bytes, out newId);
         }
 
-    #if NET6_0_OR_GREATER
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        static void FromSequentialByteSpan(ReadOnlySpan<byte> span, out NewId newId)
+        static void FromByteArray(ReadOnlySpan<byte> bytes, out NewId newId)
         {
-            Debug.Assert(span.Length == Unsafe.SizeOf<Guid>());
-            Debug.Assert(Ssse3.IsSupported && BitConverter.IsLittleEndian);
+        #if NET6_0_OR_GREATER
+            if (Ssse3.IsSupported && BitConverter.IsLittleEndian)
+            {
+                var vector = MemoryMarshal.Read<Vector128<byte>>(bytes);
+                var shuffle = Vector128.Create((byte)13, 12, 11, 10, 9, 8, 15, 14, 4, 5, 6, 7, 1, 0, 2, 3);
+                Vector128<byte> result = Ssse3.Shuffle(vector, shuffle);
+                newId = Unsafe.As<Vector128<byte>, NewId>(ref result);
+                return;
+            }
+        #endif
+            var a = (bytes[10] << 24) | (bytes[11] << 16) | (bytes[12] << 8) | bytes[13];
+            var b = (bytes[14] << 24) | (bytes[15] << 16) | (bytes[8] << 8) | bytes[9];
+            var c = (bytes[7] << 24) | (bytes[6] << 16) | (bytes[5] << 8) | bytes[4];
+            var d = (bytes[3] << 24) | (bytes[2] << 16) | (bytes[0] << 8) | bytes[1];
 
-            var vector = MemoryMarshal.Read<Vector128<byte>>(span);
-            var shuffle = Vector128.Create((byte)0, 1, 2, 3, 6, 7, 4, 5, 11, 10, 9, 8, 15, 14, 13, 12);
-            Vector128<byte> result = Ssse3.Shuffle(vector, shuffle);
-            newId = Unsafe.As<Vector128<byte>, NewId>(ref result);
+            newId = new NewId(a, b, c, d);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        static void FromSequentialByteArray(in Guid guid, out NewId newId)
+        {
+            Span<byte> bytes = stackalloc byte[16];
+            guid.TryWriteBytes(bytes);
+        #if NET6_0_OR_GREATER
+            if (Ssse3.IsSupported && BitConverter.IsLittleEndian)
+            {
+                var vector = MemoryMarshal.Read<Vector128<byte>>(bytes);
+                var shuffle = Vector128.Create((byte)0, 1, 2, 3, 6, 7, 4, 5, 11, 10, 9, 8, 15, 14, 13, 12);
+                Vector128<byte> result = Ssse3.Shuffle(vector, shuffle);
+                newId = Unsafe.As<Vector128<byte>, NewId>(ref result);
+                return;
+            }
+        #endif
+            var a = (bytes[3] << 24) | (bytes[2] << 16) | (bytes[1] << 8) | bytes[0];
+            var b = (bytes[5] << 24) | (bytes[4] << 16) | (bytes[7] << 8) | bytes[6];
+            var c = (bytes[8] << 24) | (bytes[9] << 16) | (bytes[10] << 8) | bytes[11];
+            var d = (bytes[12] << 24) | (bytes[13] << 16) | (bytes[14] << 8) | bytes[15];
+
+            newId = new NewId(a, b, c, d);
+        }
+    #else
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        static void FromGuid(in Guid guid, out NewId newId)
+        {
+            FromByteArray(guid.ToByteArray(), out newId);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        static void FromByteArray(in byte[] bytes, out NewId newId)
+        {
+            var a = (bytes[10] << 24) | (bytes[11] << 16) | (bytes[12] << 8) | bytes[13];
+            var b = (bytes[14] << 24) | (bytes[15] << 16) | (bytes[8] << 8) | bytes[9];
+            var c = (bytes[7] << 24) | (bytes[6] << 16) | (bytes[5] << 8) | bytes[4];
+            var d = (bytes[3] << 24) | (bytes[2] << 16) | (bytes[0] << 8) | bytes[1];
+
+            newId = new NewId(a, b, c, d);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        static void FromSequentialByteArray(in Guid guid, out NewId newId)
+        {
+            var bytes = guid.ToByteArray();
+            var a = (bytes[3] << 24) | (bytes[2] << 16) | (bytes[1] << 8) | bytes[0];
+            var b = (bytes[5] << 24) | (bytes[4] << 16) | (bytes[7] << 8) | bytes[6];
+            var c = (bytes[8] << 24) | (bytes[9] << 16) | (bytes[10] << 8) | bytes[11];
+            var d = (bytes[12] << 24) | (bytes[13] << 16) | (bytes[14] << 8) | bytes[15];
+
+            newId = new NewId(a, b, c, d);
         }
     #endif
     }