--- a/tests/MassTransit.BenchmarkConsole/MassTransit.BenchmarkConsole.csproj
+++ b/tests/MassTransit.BenchmarkConsole/MassTransit.BenchmarkConsole.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFrameworks>net6.0;net7.0;netcoreapp3.1</TargetFrameworks>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <PackageId>MassTransit.BenchmarkConsole</PackageId>
     <OutputType>Exe</OutputType>