@TimothyMakkison 2023-02-16
> @TimothyMakkison Tim, would you be interested in commenting on some of these tweaks or make some suggestions?

Would love to! This PR is incomplete so I'll make a couple of suggestions. It looks like this PR will remove the allocation for .Net 6 when intrinsics aren't supported. 

 - Should `#if NETSTANDARD2_1_OR_GREATER` be used for the span code. This way all versions with [span support](https://learn.microsoft.com/en-us/dotnet/api/system.span-1?view=net-7.0#applies-to) would get the benefit.
 - If you fancy microbenchmarking I'm pretty sure `FromByteArray` and `FromSequentialByteArray` could have some minor performance gains using byte swapping and `BinaryPrimitives.WriteInt32BigEndian`.
 - If you want to make a breaking change consider adding `GetFormatterArray(Span<byte> bytes)` and `GetSequentialFormatterArray(Span<byte> bytes)` and changing `INewIdFormatter.Format` to `string Format(ReadOnlySpan<byte> bytes)`. This should remove all but the string allocation from the `ToString` methods.
- Your method of converting the Guid to bytes is better than my hacky version. Why was the in keyword removed from `NewId FromGuid(Guid guid)`, is this faster? `MemoryMarshal.TryWrite(bytes, ref Unsafe.AsRef(in guid))` should produce the same code while passing `Guid` ByRef. [SharpLap](https://sharplab.io/#v2:C4LghgzgtgPgAgJgIwFgBQcAMACOSB0ASgK4B2wAllAKb4DCA9lAA4UA21ATgMpcBuFAMbUIAbnRZcBEuSq0AkuS4NmvTgOFj0EgMy4E2OtgDe6bOdx64AFmwBZABQBKE2Yvu+YTtgDm2ALzYAOLEFAAm+ABy1ADuIeHO4mjuKanYbqkA9JmKAOqcFMDUAEIAnkUOPk5JadjZ+YUl5dSV1RnYAL7t7XBWtooAKgxlFRSkwaFhvpNOpslp3MxgpAA88WEAfNgQS+OBEMBgggDWYGxsDIIT4QDaSAC6NWk7yzeY9wHT4U+pi8srACNmlsgUU/nt7NQoAxOKU7F4IAALM70SDANaTAA02FB1A2DhepDa83MXRJ6XJllwtiGIxa6y+YVm7Xc4Ix4S2hM+ByOp3Ol2uYTujxZFkJbw+gR8kx+KTZuJBzXBnzsUJhcIRyLYqIO7LC2IVBN2xPcZPcLN61OweQKRTpDjGgsZzMpFnlwJxzQg3MOJzOFyuuLuADYRa7UqrobD4ZwkSiBrCGhVcRBsZxqAAzbAAVVIEDAGdoAEEIIRMw7xtLwk4TRYzRYen1sEmmhUGVWmXNau6ioqit79r6+QHPUUQ2Hau5I+qY3HtQnSi2HCm05nnbKyR0gA===)

I dug around my branches and found the following change. It could do with using your Span methods for the non intrinsics path. No idea why I never submitted it 😆 

```C#
public NewId(in byte[] bytes)
{
    if (bytes == null)
        throw new ArgumentNullException(nameof(bytes));
    if (bytes.Length != 16)
        throw new ArgumentException("Exactly 16 bytes expected", nameof(bytes));

#if NET6_0_OR_GREATER
    if (Ssse3.IsSupported && BitConverter.IsLittleEndian)
    {
        FromByteSpan(bytes, out this);
        return;
    }
#endif

    FromByteArray(bytes, out _a, out _b, out _c, out _d);
}

public NewId(in string value)
{
    if (string.IsNullOrEmpty(value))
        throw new ArgumentException("must not be null or empty", nameof(value));

    var guid = new Guid(value);

#if NET6_0_OR_GREATER
    if (Ssse3.IsSupported && BitConverter.IsLittleEndian)
    {
        Span<byte> span = stackalloc byte[16];
        guid.TryWriteBytes(span);
        FromByteSpan(span, out this);
        return;
    }
#endif

    var bytes = guid.ToByteArray();
    FromByteArray(bytes, out _a, out _b, out _c, out _d);
}
```

#### Before
|        Method |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|-------------- |---------:|---------:|---------:|-------:|----------:|
| NewFromString | 67.45 ns | 1.014 ns | 0.899 ns | 0.0056 |      40 B |
|  NewFromBytes | 13.15 ns | 0.034 ns | 0.028 ns |      - |         - |

#### After
|        Method |      Mean |     Error |    StdDev |    Median | Allocated |
|-------------- |----------:|----------:|----------:|----------:|----------:|
| NewFromString | 54.606 ns | 1.6539 ns | 4.4147 ns | 52.742 ns |         - |
|  NewFromBytes |  3.597 ns | 0.0566 ns | 0.0473 ns |  3.588 ns |         - |