@NooNameR 2023-02-16
@TimothyMakkison I was trying to convince @phatboyg to use https://www.nuget.org/packages/System.Memory for netstandard/framework this will avoid those ifs, but he is stubborn