@NooNameR 2023-02-16
I do not have ssse3 on my M1, so just posting net6/7

### Before

|             Method | EnvironmentVariables |       Runtime |      Mean |     Error |    StdDev | Allocated |
|------------------- |--------------------- |-------------- |----------:|----------:|----------:|----------:|
|             ToGuid |                Empty |      .NET 6.0 |  2.058 ns | 0.0126 ns | 0.0112 ns |         - |
|   ToSequentialGuid |                Empty |      .NET 6.0 |  2.129 ns | 0.0077 ns | 0.0072 ns |         - |
|        ToByteArray |                Empty |      .NET 6.0 |  7.094 ns | 0.0597 ns | 0.0559 ns |      40 B |
|           FromGuid |                Empty |      .NET 6.0 | 11.164 ns | 0.1153 ns | 0.1079 ns |      40 B |
| FromSequentialGuid |                Empty |      .NET 6.0 | 11.632 ns | 0.1202 ns | 0.1124 ns |      40 B |
|      NewIdToString |                Empty |      .NET 6.0 | 57.625 ns | 0.5468 ns | 0.5114 ns |      96 B |
|             ToGuid |                Empty |      .NET 7.0 |  1.884 ns | 0.0111 ns | 0.0099 ns |         - |
|   ToSequentialGuid |                Empty |      .NET 7.0 |  1.788 ns | 0.0104 ns | 0.0087 ns |         - |
|        ToByteArray |                Empty |      .NET 7.0 |  6.311 ns | 0.0787 ns | 0.0736 ns |      40 B |
|           FromGuid |                Empty |      .NET 7.0 | 10.282 ns | 0.0289 ns | 0.0242 ns |      40 B |
| FromSequentialGuid |                Empty |      .NET 7.0 | 10.814 ns | 0.0163 ns | 0.0136 ns |      40 B |
|      NewIdToString |                Empty |      .NET 7.0 | 46.821 ns | 0.0444 ns | 0.0394 ns |      96 B |

### After

|             Method | EnvironmentVariables |       Runtime |      Mean |     Error |    StdDev | Allocated |
|------------------- |--------------------- |-------------- |----------:|----------:|----------:|----------:|
|             ToGuid |                Empty |      .NET 6.0 |  2.069 ns | 0.0133 ns | 0.0125 ns |         - |
|   ToSequentialGuid |                Empty |      .NET 6.0 |  2.108 ns | 0.0116 ns | 0.0109 ns |         - |
|        ToByteArray |                Empty |      .NET 6.0 |  6.837 ns | 0.0259 ns | 0.0243 ns |      40 B |
|           FromGuid |                Empty |      .NET 6.0 |  6.165 ns | 0.0147 ns | 0.0137 ns |         - |
| FromSequentialGuid |                Empty |      .NET 6.0 |  6.104 ns | 0.0157 ns | 0.0139 ns |         - |
|      NewIdToString |                Empty |      .NET 6.0 | 57.516 ns | 0.1745 ns | 0.1632 ns |      96 B |
|             ToGuid |                Empty |      .NET 7.0 |  1.858 ns | 0.0151 ns | 0.0134 ns |         - |
|   ToSequentialGuid |                Empty |      .NET 7.0 |  1.781 ns | 0.0083 ns | 0.0077 ns |         - |
|        ToByteArray |                Empty |      .NET 7.0 |  6.256 ns | 0.0452 ns | 0.0401 ns |      40 B |
|           FromGuid |                Empty |      .NET 7.0 |  6.021 ns | 0.0188 ns | 0.0167 ns |         - |
| FromSequentialGuid |                Empty |      .NET 7.0 |  6.066 ns | 0.0069 ns | 0.0058 ns |         - |
|      NewIdToString |                Empty |      .NET 7.0 | 46.978 ns | 0.1003 ns | 0.0938 ns |      96 B |